\name{ggsave.emf}
\alias{ggsave.emf}
\title{Save a ggplot object as an emf file}
\usage{
  ggsave.emf(plot, filename, ...)
}
\arguments{
  \item{plot}{The ggplot object to write.}

  \item{filename}{The name of the file.}

  \item{...}{The additional arguments to pass on to
  \code{\link{emf}}.}
}
\description{
  Save a ggplot object as an emf file
}

